remove unnecessary brackets & revert em-dashboard-pom.xml & move RocketmqProperties to core.dto
